Fixed /del

+++ bot/db/migrations/004_add_thread_to_subs.sql
@@ -1,7 +1,7 @@
 -- +goose Up
 
 ALTER TABLE subs
-ADD COLUMN thread INTEGER NOT NULL DEFAULT 0;
+ADD COLUMN thread BIGINT NOT NULL DEFAULT 0;
 
 -- +goose Down
